refactor: extract student info constants
