+++ src/constants/student-info.ts
@@ -1,0 +1,25 @@
+export const STUDENT_STATUSES = [
+    'At The University of Adelaide',
+    'At another university',
+    'No',
+] as const;
+export const DEGREES = [
+    'Bachelor of Computer Science or Bachelor of Computer Science (Advanced)',
+    'Bachelor of Maths & Computer Science',
+    'Bachelor of IT',
+    'Bachelor of Software Engineering',
+    'Honours, Computer Science',
+    'Masters (Coursework), Computer Science',
+    'Masters/PhD (Research), Computer Science',
+    'Other',
+] as const;
+export const AGE_BRACKETS = [
+    'Under 20',
+    '20-24',
+    '25-29',
+    '30-34',
+    'Over 34',
+    'Prefer not to say',
+] as const;
+export const GENDERS = ['Male', 'Female', 'Other', 'Prefer not to say'] as const;
+export const STUDENT_TYPES = ['Domestic', 'International'] as const;
